worker/requirements.txt
celery>=5.4.0
redis>=5.0.7
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.7.0
jsonschema>=4.23.0
openai>=1.30.0
anthropic>=0.26.0
google-generativeai>=0.5.0
twilio>=9.0.0
resend>=0.8.0
pillow>=10.3.0
instaloader>=4.10.0
google-search-results>=2.4.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
